requirements.txt
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
bitsandbytes==0.49.1
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
datasets==4.5.0
dill==0.4.0
entmax==1.3
filelock==3.20.0
frozenlist==1.8.0
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
jsonargparse==3.13.1
lightning-utilities==0.15.2
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.18
networkx==3.6.1
numpy==1.26.4
pandas==3.0.0
peft==0.18.1
pillow==12.0.0
platformdirs==4.5.1
portalocker==3.2.0
propcache==0.4.1
protobuf==4.25.8
psutil==7.2.2
pyarrow==23.0.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
pytorch-lightning==2.6.1
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
rich==14.3.2
sacrebleu==2.6.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
sentence-transformers==5.2.2
sentencepiece==0.2.1
sentry-sdk==2.52.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sympy==1.13.1
tabulate==0.9.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchmetrics==0.10.3
torchvision==0.20.1+cu121
tqdm==4.67.3
transformers==4.57.6
triton==3.1.0
trl==0.28.0
typer==0.23.0
typer-slim==0.23.0
typing-inspection==0.4.2
typing_extensions==4.15.0
unbabel-comet==2.2.7
urllib3==2.6.3
wandb==0.24.2
xxhash==3.6.0
yarl==1.22.0
